+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Adding the `run` command.
 * Adding the `build` command.
+* Adding the `--minify` option.
+* Improving UX.
 
 ## 0.6.0